requirements-test.txt
black
codecov
coverage
flake8
flask-debugtoolbar
flask-shell-ipython
gitchangelog
ipdb
isort
mkdocs
mypy
pre-commit
pytest
pytest-cov
pytest-flask
python-dotenv
